requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
asttokens==3.0.0
attrs==25.4.0
billiard==4.2.1
CacheControl==0.14.3
cachetools==5.5.2
celery==5.5.3
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
cryptography==45.0.6
decorator==5.2.1
distro==1.9.0
dj-database-url==3.0.1
Django==5.2.1
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-cron==0.6.0
django-js-asset==3.1.2
django-mptt==0.18.0
django-phonenumber-field==8.1.0
django-timezone-field==7.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
drf-spectacular==0.28.0
drf-yasg==1.21.10
executing==2.2.0
firebase_admin==7.1.0
# google-ai-generativelanguage==0.6.15
# google-api-core==2.25.1
# google-api-python-client==2.176.0
# google-auth==2.40.3
# google-cloud-core==2.4.3
# google-cloud-firestore==2.21.0
# google-cloud-storage==3.2.0
# google-crc32c==1.7.1
# google-generativeai==0.8.5
# google-resumable-media==2.7.2
# googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
inflection==0.5.1
ipython==4.0.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jiter==0.10.0
jsonschema==4.17.3
kombu==5.5.4
matplotlib-inline==0.1.7
msgpack==1.1.1
numpy==2.2.3
openai==1.97.0
opencv-python==4.11.0.86
packaging==25.0
parso==0.8.4
phonenumbers==9.0.10
pickleshare==0.7.5
pillow==11.3.0
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==6.31.1
psycopg2-binary==2.9.10
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.3
pyrsistent==0.20.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
requests==2.32.4
rsa==4.9.1
simplegeneric==0.8.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
stack-data==0.6.3
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
